Add a create table helper
